New version of BoxologyPlugin
In this version, most of the problem are detected, and now it can support better the complex pattern. Also this version is base on rules which are mentioned in the report, and shows the rule after each validation. 
If a pattern is invalid, you can put the pointer on nodes and see the problem as tooltip.